intel_rdt: refactor pids monitoring code. Extract pids config to seperate file. Extract pids config tests to seperate file.
